+++ pages/components/index.ts
@@ -1,0 +1,6 @@
+export { default as AnimatedTextSection } from './AnimatedTextSection';
+export { default as GridSection } from './GridSection';
+export { default as IntroSection } from './IntroSection';
+export { default as NewsSection } from './NewsSection';
+export { default as ProjectSection } from './ProjectSection';
+export { default as SponsorSection } from './SponsorSection';
